backend/requirements.txt
fastapi
uvicorn
google-generativeai
anthropic
pypdf
python-dotenv
python-multipart
